test task done

+++ src/app/todo-list/todo-list.component.ts
@@ -1,19 +1,28 @@
-import { Component, OnInit, AfterViewInit, ViewChild } from '@angular/core';
-import { Input } from '@angular/core';
-
+import {
+  AfterViewInit,
+  Component,
+  OnDestroy,
+  OnInit,
+  ViewChild,
+} from '@angular/core';
+import { Observable, Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+import { MatDialog } from '@angular/material/dialog';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
-import { Subject } from 'rxjs';
 
-import { Todo } from '../shared/models';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
+import { TodoFormComponent } from '../todo-form/todo-form.component';
 import { TodosStateService } from '../todos-state.service';
+import { INITIAL_TODOS } from '../shared/constants';
+import { Todo } from '../shared/models';
 
 @Component({
   selector: 'app-todo-list',
   templateUrl: './todo-list.component.html',
   styleUrls: ['./todo-list.component.scss'],
 })
-export class TodoListComponent implements OnInit, AfterViewInit {
+export class TodoListComponent implements OnInit, AfterViewInit, OnDestroy {
   displayedColumns: string[] = [
     'name',
     'description',
@@ -22,42 +31,74 @@ export class TodoListComponent implements OnInit, AfterViewInit {
     'expiryDate',
     'actions',
   ];
-  @Input() todoList: Todo[] = [];
-  // @ts-ignore
-  dataSource: MatTableDataSource<Todo>;
+  dataSource: MatTableDataSource<Todo> | undefined;
   todoList$ = this.todosStateService.todos$;
-  selectedTodo: Subject<Todo> = new Subject<Todo>();
+  onDestroy: Subject<null> = new Subject<null>();
 
-  // @ts-ignore
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatSort) sort: MatSort | undefined;
 
-  constructor(private todosStateService: TodosStateService) {}
+  constructor(
+    private todosStateService: TodosStateService,
+    public dialog: MatDialog
+  ) {}
 
   ngOnInit(): void {
-    // this.dataSource = new MatTableDataSource(this.todoList);
-    this.todoList$.subscribe((list) => {
+    this.setInitialTodos();
+    this.todoList$.pipe(takeUntil(this.onDestroy)).subscribe((list) => {
       this.dataSource = new MatTableDataSource(list);
     });
-    const newTodo: Todo = {
-      name: 'Test 1',
-      description: 'Empty',
-      completed: false,
-      favourite: true,
-      expiryDate: new Date(Date.now() + 3600000),
-    };
-    this.todosStateService.addTodo(newTodo);
   }
 
   ngAfterViewInit(): void {
-    // @ts-ignore
-    this.dataSource.sort = this.sort;
+    this.todoList$.pipe(takeUntil(this.onDestroy)).subscribe(() => {
+      if (this.dataSource && this.sort) {
+        this.dataSource.sort = this.sort;
+      }
+    });
+  }
+
+  ngOnDestroy(): void {
+    this.onDestroy.next();
+  }
+
+  setInitialTodos(): void {
+    INITIAL_TODOS.forEach((todo) => {
+      this.todosStateService.addTodo(todo);
+    });
+  }
+
+  addTodo(): void {
+    this.dialog.open(TodoFormComponent, {
+      width: '320px',
+      data: { todo: null },
+    });
   }
 
   editTodo(todo: Todo): void {
-    this.selectedTodo.next(todo);
+    this.dialog.open(TodoFormComponent, {
+      width: '320px',
+      data: { todo },
+    });
   }
 
-  deleteTodo(id: number): void {
-    this.todosStateService.removeTodo(id);
+  deleteTodo(id: string): void {
+    this.openConfirmDialog().subscribe((result) => {
+      if (result) {
+        this.todosStateService.removeTodo(id);
+      }
+    });
+  }
+
+  openConfirmDialog(): Observable<boolean> {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '320px',
+      data: { text: 'Are you sure you want to delete this todo record?' },
+    });
+
+    return dialogRef.afterClosed();
+  }
+
+  getDateExpiredClass(expiryDate: Date): string {
+    return +expiryDate < +new Date() ? 'date-expired' : '';
   }
 }
